requirements.txt
PyQt6>=6.6.0
openai-whisper>=20231117
scikit-learn>=1.0.0
numpy>=1.20.0
